+++ inst/NEWS.md
@@ -4,14 +4,13 @@ __with Gating and Expert Network Covariates__
 =======================================================
 
 ## MoEClust v1.2.0 - (_4<sup>th</sup> release [minor update]: 2018-08-25_)
-### New Features & Improvements
+### New Features, Improvements, Bug Fixes & Miscellaneous Edits
 * New `MoE_control` arg. `algo` allows model fitting using the `"EM"` or `"CEM"` algorithm:  
     * Related new function `MoE_cstep` added.
     * Extra `algo` option `"cemEM"` allows running EM starting from convergence of CEM.
 * New `MoE_control` arg. `nstarts` allows for multiple random starts when `init.z="random"`.
+* New `MoE_control` arg. `tau0` provides another means of initialising the noise component.
 * If `clustMD` is invoked for initialisation, models are now run more quickly in parallel.
-
-### Bug Fixes & Miscellaneous Edits
 * Fixed bug in checking for strictly increasing log-likelihood estimates.
 
 ## MoEClust v1.2.0 - (_3<sup>rd</sup> release [minor update]: 2018-08-24_)